Initial upload (Doesn't run)
Initial upload for the project.
It worked when I turned in the assignment. 
Something broke throughout the years and I can't get the code to run without significant effort.
Source code seems to be good, and VSCode doesn't give out errors.
Requires Java and JavaFX. Otherwise, the dependencies are in the files.